+++ changelog.md
@@ -1,3 +1,11 @@
+# Version 2025.12.2 (2025-12-01)
+
+## What's Changed
+
+- Add HubProtocol and WeekProfileProtocol to interfaces
+- Rename regaid to rega_id
+- Update documentation
+
 # Version 2025.12.1 (2025-12-01)
 
 ## What's Changed